miad, wue ham

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: mresch <mresch@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/04 17:34:21 by mresch            #+#    #+#             */
-/*   Updated: 2023/09/06 12:01:48 by mresch           ###   ########.fr       */
+/*   Updated: 2023/09/07 16:48:39 by mresch           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int		i;
-	char	*p;
+	size_t		i;
+	char		*p;
 
 	i = 0;
 	p = s;
