+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mresch <mresch@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 14:25:17 by mresch            #+#    #+#             */
-/*   Updated: 2023/09/05 15:38:23 by mresch           ###   ########.fr       */
+/*   Updated: 2023/09/07 16:47:26 by mresch           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] != 0 && i < n)
